rename: 컴포넌트 파일 구조 컨벤션 변경으로 인한 리프 컴포넌트 경로 및 파일명 변경 (#102)
